Optimization Pass:
* Optimized the PathHelper.GetProcessedPathContents method.  The optimizations made the sorting step unnecessary and improved speed by a quarter on 1.25 million records, stripping five minutes off.

* Added ToString overides to the RecordActionContainer and ScanEvent classes.

* Converted the DiskScanEventWrapper to report its final information via the HDSLQueryOutcome class system.

* TotalFiles and TotalDirectories on the PathSetData class are now long integers.

* PathHelper.GetDependencyCount no longer needs to hit the disk to do its job, and there is now an overload that takes a simple path and whether or not it is a file or directory as parameters.

+++ Harddrive-Library/IO/Disk/PathSetData.cs
@@ -19,12 +19,12 @@ namespace HDDL.IO.Disk
         /// <summary>
         /// The total number of files to scan
         /// </summary>
-        public int TotalFiles { get; set; }
+        public long TotalFiles { get; set; }
 
         /// <summary>
         /// The total number of directories
         /// </summary>
-        public int TotalDirectories { get; set; }
+        public long TotalDirectories { get; set; }
 
         /// <summary>
         /// Create a PathSetData
